+++ packages/dappmanager/src/modules/https-portal/index.ts
@@ -9,8 +9,9 @@ import params from "../../params";
 import { getExternalNetworkAlias } from "../../domains";
 import { PackageContainer, HttpsPortalMapping } from "../../types";
 import { HttpsPortalApiClient } from "./apiClient";
-import { addNetworkAliasCompose, removeNetworkAliasCompose } from "./utils";
 import { ComposeEditor } from "../compose/editor";
+import { addNetworkAliasCompose } from "./utils/addNetworkAliasCompose";
+import { removeNetworkAliasCompose } from "./utils/removeNetworkAliasCompose";
 export { addAliasToRunningContainersMigration } from "./migration";
 export { HttpsPortalApiClient };
 export { getExposableServices } from "./exposable";
