+++ packages/dappmanager/src/modules/https-portal/utils/addNetworkAliasCompose.ts
@@ -1,0 +1,13 @@
+import { PackageContainer } from "../../../types";
+import { ComposeFileEditor } from "../../compose/editor";
+
+export function addNetworkAliasCompose(
+  container: PackageContainer,
+  networkName: string,
+  aliases: string[]
+): void {
+  const compose = new ComposeFileEditor(container.dnpName, container.isCore);
+  const composeService = compose.services()[container.serviceName];
+  composeService.addNetwork(networkName, { aliases });
+  compose.write();
+}
